+++ Codes/Hotel_Client/src/po/HotelPO.java
@@ -18,7 +18,7 @@ public class HotelPO extends BriefHotelInfoPO {
 	private HashMap<RoomType, Integer> roomTypeAndNums;
 	private HashMap<String, String> comments;
 	
-	public HotelPO(String hotelName, String businessDistrict, String hotelAddress, int starLevel, int mark,
+	public HotelPO(String hotelName, String businessDistrict, String hotelAddress, int starLevel, float mark,
 			String briefIntroduction, String facilityAndService, HashMap<RoomType, Integer> roomTypeAndPrice,HashMap<RoomType, Integer> roomTypeAndNums,
 			HashMap<String, String> comments) {
 		super(hotelName, businessDistrict, hotelAddress, starLevel, mark);
